Update quence.lua
I added parameter options so that each track can be independently selected to output to midi, audio, midi & audio, crow JF or crow CV. I think it's working, but @justmat please test the crow options when you get a chance.

I wanted to add the ability to assign specific MIDI devices and channels to each track but I haven't thought of an elegant way to add this to the parameter menu without exploding the menu. With 4 tracks, 4 devices and 16 channels there are a lot of combinations. I'll give it some more thought, but please let me know if anyone has any ideas.

MIDI channels are still based on the track number. So if track 1 is MIDI and track 2 is Audio, track 3 will be channel 3 if MIDI is selected.

One other minor tweak -- I added two variables, pxcurve1 and pxcurve2 which control the randomization curves for notes and rests. Previously these were set at a constant. They are currently not changeable parameters, but can be easily tweaked in the code.