improvement to transposition
This changes addresses issue #35. 

Previously, when a transposition would result in a note going above/below the limits of the scale vector, the note looped around to the bottom/top of the scale. This sounded bad, but preserved the positioning of the sequence so that if you transposed in the opposite direction you could recover the original sequence.

Now, a transposition above/below the limit of the scale vector will use the highest/lowest note within the scale vector for playing notes. But the relative positioning will be preserved and the original sequence can be recovered by transposing in the opposite direction.